email-notification/requirements.txt
Flask
python-dotenv
requests
Flask-CORS